Split CodeBlock and NotesPopup into separate components
